+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changes
 0.9 (unreleased)
 ----------------
 
+ - launchpad #1196310: provide a "post buildout steps" for functional testing
  - launchpad #1196308: provide a "post buildout steps" for nose testing
  - launchpad #1142994: url rewrites for vcs polling
  - launchpad #1154673: treeStableTimer scheduler parameter now